fix: search dialog and edit link
