+++ composables/useEditLink.ts
@@ -1,13 +1,21 @@
-export function useEditLink() {
-  const { page } = useContent();
+import type { DocEnCollectionItem } from "@nuxt/content";
+
+export function useEditLink(page: DocEnCollectionItem) {
+  const { locale } = useI18n();
   const { enable, pattern, text, icon, placement } = useConfig().value.main.editLink;
 
   const url = computed(
-    () => pattern.replace(/:path/g, page.value._file ?? ''),
+    () => {
+      let path = page?.path ?? '';
+      if (path.startsWith(`/${locale.value}/`)) {
+        path = path.replace(`/${locale.value}/`, '/');
+      }
+      return pattern.replace(/:path/g, path)
+    },
   );
 
   const enabled = computed(
-    () => enable && page.value?.editLink !== false && page.value?._file && url.value !== '',
+    () => enable && page?.meta?.editLink !== false && page?.path && url.value !== '',
   );
 
   const enabledToc = computed(
